Support sourced shell snippets as well
The former script has been split into a helper executable and two
substitutes.

The first substitute is intended to replace executable scripts of any
kind and can be symlinked to the original location directly.

The other substitute is a template for a shell script snippet and must
be customized with the absolute path of the original snippet in order to
work reliably.